Added testenv port to deploy, acceptance and capacity scripts. Added  to deploy script

+++ deploy.sh
@@ -1,8 +1,10 @@
 #!/bin/bash
 if [ -z "$1" ]; then echo "Empty parameter for testenv IP Address"; exit -1; fi
+if [ -z "$2" ]; then echo "Empty parameter for testenv PORT"; exit -1; fi
+if [ -z "$3" ]; then echo "Empty parameter for last successful commit"; exit -1; fi
 
 echo Executing remote...
 ssh vagrant@"$1" "docker stop tictactoe;
-							docker rm tictactoe;
-							docker pull ggjonsson/tictactoe;
-							docker run -p 9000:8080 --name tictactoe -d -e "NODE_ENV=production" ggjonsson/tictactoe"
+				docker rm tictactoe;
+				docker pull ggjonsson/tictactoe"$2";
+				docker run -p 9000:"$2" --name tictactoe -d -e "NODE_ENV=production" ggjonsson/tictactoe"$2""
